# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RejeXion API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /growth-notes/{id}:
    delete:
      operationId: deleteGrowthNote
      tags: [growth-notes]
      summary: Delete a growth note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Growth note deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /entries/{entryId}/growth-notes:
    get:
      operationId: listGrowthNotes
      tags: [growth-notes]
      summary: List growth notes for an entry
      parameters:
        - name: entryId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of growth notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GrowthNote"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createGrowthNote
      tags: [growth-notes]
      summary: Add a growth note to an entry
      parameters:
        - name: entryId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGrowthNoteBody"
      responses:
        "201":
          description: Growth note created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GrowthNote"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    GrowthNote:
      type: object
      properties:
        id:
          type: string
        entryId:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - entryId
        - content
        - createdAt
    CreateGrowthNoteBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content
...