# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RejeXion API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Register a new user
 */
export const registerBodyPasswordMin = 8;

export const RegisterBody = zod.object({
  email: zod.string().email(),
  password: zod.string().min(registerBodyPasswordMin),
  name: zod.string(),
});

/**
 * @summary Login
 */
export const LoginBody = zod.object({
  email: zod.string().email(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  user: zod.object({
    id: zod.string(),
    email: zod.string(),
    name: zod.string(),
    createdAt: zod.coerce.date(),
  }),
});

/**
 * @summary Logout
 */
export const LogoutResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Get current user
 */
export const GetMeResponse = zod.object({
  id: zod.string(),
  email: zod.string(),
  name: zod.string(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Request password reset
 */
export const ForgotPasswordBody = zod.object({
  email: zod.string().email(),
});

export const ForgotPasswordResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Reset password with token
 */
export const resetPasswordBodyPasswordMin = 8;

export const ResetPasswordBody = zod.object({
  token: zod.string(),
  password: zod.string().min(resetPasswordBodyPasswordMin),
});

export const ResetPasswordResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Get user profile
 */
export const GetProfileResponse = zod.object({
  id: zod.string(),
  email: zod.string(),
  name: zod.string(),
  totalPoints: zod.number(),
  entryCount: zod.number(),
  growthNoteCount: zod.number(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Get prompt templates by category
 */
export const GetPromptsQueryParams = zod.object({
  category: zod.enum(["academic", "career"]),
});

export const GetPromptsResponseItem = zod.object({
  id: zod.string(),
  category: zod.enum(["academic", "career"]),
  promptText: zod.string(),
  orderIndex: zod.number(),
});
export const GetPromptsResponse = zod.array(GetPromptsResponseItem);

/**
 * @summary List user entries
 */
export const ListEntriesQueryParams = zod.object({
  category: zod.enum(["academic", "career", "all"]).optional(),
});

export const ListEntriesResponseItem = zod.object({
  id: zod.string(),
  category: zod.enum(["academic", "career"]),
  title: zod.string(),
  rejectionDate: zod.coerce.date(),
  description: zod.string(),
  growthNoteCount: zod.number(),
  createdAt: zod.coerce.date(),
});
export const ListEntriesResponse = zod.array(ListEntriesResponseItem);

/**
 * @summary Create a new rejection entry
 */
export const CreateEntryBody = zod.object({
  category: zod.enum(["academic", "career"]),
  title: zod.string(),
  rejectionDate: zod.coerce.date(),
  description: zod.string(),
  responses: zod.array(
    zod.object({
      promptId: zod.string(),
      responseText: zod.string(),
    }),
  ),
});

/**
 * @summary Get a specific entry
 */
export const GetEntryParams = zod.object({
  id: zod.coerce.string(),
});

export const GetEntryResponse = zod.object({
  id: zod.string(),
  category: zod.enum(["academic", "career"]),
  title: zod.string(),
  rejectionDate: zod.coerce.date(),
  description: zod.string(),
  responses: zod.array(
    zod.object({
      id: zod.string(),
      promptId: zod.string(),
      promptText: zod.string(),
      responseText: zod.string(),
    }),
  ),
  growthNotes: zod.array(
    zod.object({
      id: zod.string(),
      entryId: zod.string(),
      content: zod.string(),
      createdAt: zod.coerce.date(),
    }),
  ),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Update an entry's basic details
 */
export const UpdateEntryParams = zod.object({
  id: zod.coerce.string(),
});

export const UpdateEntryBody = zod.object({
  title: zod.string().optional(),
  rejectionDate: zod.coerce.date().optional(),
  description: zod.string().optional(),
});

export const UpdateEntryResponse = zod.object({
  id: zod.string(),
  category: zod.enum(["academic", "career"]),
  title: zod.string(),
  rejectionDate: zod.coerce.date(),
  description: zod.string(),
  growthNoteCount: zod.number(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete an entry
 */
export const DeleteEntryParams = zod.object({
  id: zod.coerce.string(),
});

/**
 * @summary Export entry as plain text
 */
export const ExportEntryParams = zod.object({
  id: zod.coerce.string(),
});

/**
 * @summary Delete a growth note
 */
export const DeleteGrowthNoteParams = zod.object({
  id: zod.coerce.string(),
});

/**
 * @summary List growth notes for an entry
 */
export const ListGrowthNotesParams = zod.object({
  entryId: zod.coerce.string(),
});

export const ListGrowthNotesResponseItem = zod.object({
  id: zod.string(),
  entryId: zod.string(),
  content: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListGrowthNotesResponse = zod.array(ListGrowthNotesResponseItem);

/**
 * @summary Add a growth note to an entry
 */
export const CreateGrowthNoteParams = zod.object({
  entryId: zod.coerce.string(),
});

export const CreateGrowthNoteBody = zod.object({
  content: zod.string(),
});

/**
 * @summary Get user points summary
 */
export const GetPointsResponse = zod.object({
  totalPoints: zod.number(),
  entryPoints: zod.number(),
  growthNotePoints: zod.number(),
  entryCount: zod.number(),
  growthNoteCount: zod.number(),
});

/**
 * @summary Get user entry statistics
 */
export const GetStatsResponse = zod.object({
  totalEntries: zod.number(),
  academicEntries: zod.number(),
  careerEntries: zod.number(),
  totalGrowthNotes: zod.number(),
  totalPoints: zod.number(),
  recentEntries: zod.array(
    zod.object({
      id: zod.string(),
      category: zod.enum(["academic", "career"]),
      title: zod.string(),
      rejectionDate: zod.coerce.date(),
      description: zod.string(),
      growthNoteCount: zod.number(),
      createdAt: zod.coerce.date(),
    }),
  ),
});